+++ .github/release-notes.md
@@ -1,2 +1,3 @@
 * Fixed error when launching in setup mode
 * Added title to dialog when picking directory
+* Added setting to open web UI when TFView is opened